Progetto con sprite semidefinitivo senza gui e cambi sprite

+++ Assets/Scripts/Wave.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Wave : MonoBehaviour {
 
 	GameObject[] childArray;
-	public int minionSpawnIn;
+	public float minionSpawnIn;
 	[HideInInspector]
 	public int activator = 0;
